Added role delete verification.

+++ api/Data/Services/Roles/RolesService.cs
@@ -1,9 +1,12 @@
 ﻿using api.Data.Entities;
 using api.Data.Repositories;
+using api.Data.Repositories.Users;
 using api.Data.Requests;
 using api.Data.Responses;
 using AutoMapper;
+using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
 
 namespace api.Data.Services.Roles
@@ -11,12 +14,14 @@ namespace api.Data.Services.Roles
     public class RolesService : IBaseService<RolesResponse, RolesRequest>
     {
         private readonly IBaseRepository<RolesEntity> _repository;
+        private readonly IUsersRepository _userRepository;
         private readonly IMapper _mapper;
 
-        public RolesService(IBaseRepository<RolesEntity> repository, IMapper mapper)
+        public RolesService(IBaseRepository<RolesEntity> repository, IMapper mapper, IUsersRepository userRepository)
         {
             _repository = repository;
             _mapper = mapper;
+            _userRepository = userRepository;
         }
 
         public List<RolesResponse> GetAll()
@@ -47,6 +52,13 @@ namespace api.Data.Services.Roles
 
         public async Task DeleteAsync(int id)
         {
+            var users = _userRepository.GetAll();
+            
+            if (users.Where(u => u.RoleId == id).Any())
+            {
+                throw new Exception("Can't delete role because it has a user using it.");
+            }
+            
             await _repository.DeleteAsync(id);
         }
     }
